RAG_Assignment/requirements.txt
anthropic
chromadb
pdfplumber
python-dotenv
tqdm
pydantic-settings
sentence-transformers
